feat: restructure folders for organization and add functionality to open builds in browser
